rename: 클래스명 DiscountEventRule을 DiscountEventRules로 수정
