+++ src/main/java/christmas/model/event/WeekdayDiscountEvent.java
@@ -1,7 +1,7 @@
 package christmas.model.event;
 
 import static christmas.util.constant.DiscountConstant.NON_DISCOUNT_COST;
-import static christmas.util.rule.DiscountEventRule.WEEKDAY_DISCOUNT_EVENT;
+import static christmas.util.rule.DiscountEventRules.WEEKDAY_DISCOUNT_EVENT;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
