+++ src/main/java/christmas/util/rule/DiscountEventRules.java
@@ -1,6 +1,6 @@
 package christmas.util.rule;
 
-public enum DiscountEventRule {
+public enum DiscountEventRules {
     WEEKDAY_DISCOUNT_EVENT("평일 할인", 2023, "dessert"),
     WEEKEND_DISCOUNT_EVENT("주말 할인", 2023, "main"),
     SPECIAL_DISCOUNT_EVENT("특별 할인", 1000, "total cost");
@@ -9,7 +9,7 @@ public enum DiscountEventRule {
     private final int discountCost;
     private final String appliedTarget;
 
-    DiscountEventRule(String eventName, int discountCost, String appliedTarget) {
+    DiscountEventRules(String eventName, int discountCost, String appliedTarget) {
         this.eventName = eventName;
         this.discountCost = discountCost;
         this.appliedTarget = appliedTarget;
